#515: Rename files to be role agnostic
